fix(L3): sparse matrices must be float32, and a proof run that found it
LightGBM's CSR path rejects integer dtypes outright -- _c_float_array raises
"Expected np.float32 or np.float64, met type(int8)". load_year() was building
int8 sparse matrices, so training would have failed on first contact.

Found by actually running a training pass rather than trusting the loader
test, which only checked shape and density. Dense stays int8 while
materialising a year (4.5x smaller than float32 at peak); the cast happens at
the sparse boundary.

Proof run, training on 2020 alone and scoring later years as held-out future:

  train 2020 (102,073 rows)   16s, peak RSS 1.19 GB
    holdout AUROC  0.9965    ECE 0.0054 -> 0.0000 after isotonic
    2021           0.9905
    2022           0.9906
    2024           0.8978

The 2024 row is the result that matters. AUROC decaying 0.99 -> 0.90 across
four years is the concept-drift exhibit LAMDA exists to produce; a model still
at 0.99 in 2024 would mean the split had leaked. Positive rate also falls from
0.454 to 0.016 over the same span, so the later years are a genuinely
different problem, not just a harder one.

The full train (2013-2022, ~860k rows, ~6 GB) is deferred: 2 GB free with the
benign corpus run holding half the cores, and OOM-killing a job 440/600
through would cost three hours.

Also hardens tools/fit_calibration.py, which now refuses on three independent
grounds rather than one. The third was not anticipated and only appeared when
run: at 640 malware against 4 benign the malware prevalence is 99.4%, so the
empirical posterior P(malware|S) exceeds 0.5 almost everywhere and s0 measures
corpus composition rather than evidence. Also guards a collapsed temperature,
which turns the sigmoid into a step function.

+++ L3/train.py
@@ -90,11 +90,16 @@ def load_year(year: int, split: str, root: Path = LAMDA_BASELINE) -> YearData | 
     y = np.asarray(table.column("label").to_numpy(), dtype=np.int8)
     families = np.asarray(table.column("family").to_pylist(), dtype=object)
 
+    # Built dense as int8 (the parquet's own type, 4.5x smaller than float32
+    # while materialising a year), then converted straight to sparse float32.
+    # LightGBM's CSR path rejects integer dtypes outright — _c_float_array
+    # raises "Expected np.float32 or np.float64, met type(int8)" — so the cast
+    # has to happen somewhere, and doing it here keeps the dense peak low.
     dense = np.empty((table.num_rows, len(feat_cols)), dtype=np.int8)
     for i, col in enumerate(feat_cols):
         dense[:, i] = table.column(col).to_numpy(zero_copy_only=False)
     del table
-    X = sparse.csr_matrix(dense)
+    X = sparse.csr_matrix(dense, dtype=np.float32)
     del dense
     return YearData(year=year, X=X, y=y, families=families)
